5th commit: admin router added

+++ routers/todos.py
@@ -8,7 +8,7 @@ from database import SessionLocal
 from .auth import get_current_user
 
 router = APIRouter(
-    tags=['Todos']
+    tags=['todos']
 )
 
 def get_db():
@@ -61,10 +61,14 @@ async def create_todo(user:user_dependency,
 
 # updating todo : endpoint
 @router.put("/todo/{todo_id}", status_code = status.HTTP_204_NO_CONTENT)
-async def update_todo(db:db_dependency,
+async def update_todo(user:user_dependency,
+                      db:db_dependency,
                       todo_id:int,
                       todo_request:TodoRequest):
-    todo_model = db.query(Todos).filter(Todos.id==todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed..')
+    todo_model = db.query(Todos).filter(Todos.id==todo_id)\
+        .filter(Todos.owner_id==user.get('id')).first()
     if todo_model is  None:
         raise HTTPException(status_code=404, detail="Todo Not Found...")
     todo_model.title = todo_request.title
@@ -76,9 +80,13 @@ async def update_todo(db:db_dependency,
 
 # deleting a todo : endpoint
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db:db_dependency,
+async def delete_todo(user:user_dependency,
+                      db:db_dependency,
                       todo_id:int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id==todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed..')
+    todo_model = db.query(Todos).filter(Todos.id==todo_id)\
+        .filter(Todos.owner_id==user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo is not found...")
     db.query(Todos).filter(Todos.id==todo_id).delete()
